netx: move legacy code into the legacy folder (1/n)
This visually simplifies taking apart what is netx today and what
is legacy netx that we will make unnecessary soon.

Part of https://github.com/ooni/probe-engine/issues/746
