+++ experiment/tor/tor_test.go
@@ -15,8 +15,8 @@ import (
 	"github.com/google/go-cmp/cmp"
 	"github.com/ooni/probe-engine/internal/handler"
 	"github.com/ooni/probe-engine/internal/mockable"
-	"github.com/ooni/probe-engine/internal/oonidatamodel"
-	"github.com/ooni/probe-engine/internal/oonitemplates"
+	"github.com/ooni/probe-engine/legacy/oonidatamodel"
+	"github.com/ooni/probe-engine/legacy/oonitemplates"
 	"github.com/ooni/probe-engine/model"
 	"github.com/ooni/probe-engine/netx/modelx"
 	"github.com/ooni/probe-engine/probeservices"
