+++ geolocate/mmdblookup_test.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/apex/log"
 	"github.com/ooni/probe-engine/geolocate"
-	"github.com/ooni/probe-engine/resources"
 	"github.com/ooni/probe-engine/model"
+	"github.com/ooni/probe-engine/resources"
 )
 
 const (
