+++ legacy/README.md
@@ -1,0 +1,3 @@
+# Package github.com/ooni/probe-engine/legacy
+
+This package contains legacy code that we will soon remove.
