+++ legacy/netx/handlers/handlers_test.go
@@ -3,7 +3,7 @@ package handlers_test
 import (
 	"testing"
 
-	"github.com/ooni/probe-engine/netx/handlers"
+	"github.com/ooni/probe-engine/legacy/netx/handlers"
 	"github.com/ooni/probe-engine/netx/modelx"
 )
