+++ legacy/netx/oldhttptransport/tracetripper.go
@@ -11,10 +11,10 @@ import (
 	"time"
 
 	"github.com/ooni/probe-engine/atomicx"
+	"github.com/ooni/probe-engine/legacy/netx/connid"
+	"github.com/ooni/probe-engine/legacy/netx/dialid"
+	"github.com/ooni/probe-engine/legacy/netx/transactionid"
 	"github.com/ooni/probe-engine/netx/errorx"
-	"github.com/ooni/probe-engine/netx/internal/connid"
-	"github.com/ooni/probe-engine/netx/internal/dialid"
-	"github.com/ooni/probe-engine/netx/internal/transactionid"
 	"github.com/ooni/probe-engine/netx/modelx"
 )
