+++ legacy/netx/resolver_internal_test.go
