+++ legacy/netx/resolver_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 	"time"
 
+	"github.com/ooni/probe-engine/legacy/netx/handlers"
 	"github.com/ooni/probe-engine/netx"
-	"github.com/ooni/probe-engine/netx/handlers"
 )
 
 func testresolverquick(t *testing.T, network, address string) {
