+++ legacy/netxlogger/netxlogger_test.go
