+++ legacy/oonidatamodel/oonidatamodel_test.go
@@ -10,7 +10,7 @@ import (
 	"testing"
 	"time"
 
-	"github.com/ooni/probe-engine/internal/oonitemplates"
+	"github.com/ooni/probe-engine/legacy/oonitemplates"
 	"github.com/ooni/probe-engine/netx/modelx"
 )
