+++ legacy/oonitemplates/oonitemplates.go
@@ -22,8 +22,8 @@ import (
 	goptlib "git.torproject.org/pluggable-transports/goptlib.git"
 	"github.com/ooni/probe-engine/atomicx"
 	"github.com/ooni/probe-engine/internal/runtimex"
-	"github.com/ooni/probe-engine/netx"
-	"github.com/ooni/probe-engine/netx/handlers"
+	"github.com/ooni/probe-engine/legacy/netx"
+	"github.com/ooni/probe-engine/legacy/netx/handlers"
 	"github.com/ooni/probe-engine/netx/modelx"
 	"gitlab.com/yawning/obfs4.git/transports"
 	obfs4base "gitlab.com/yawning/obfs4.git/transports/base"
