+++ netx/archival/archival_test.go
@@ -13,11 +13,11 @@ import (
 	"github.com/apex/log"
 	"github.com/google/go-cmp/cmp"
 	"github.com/gorilla/websocket"
-	"github.com/ooni/probe-engine/resources"
 	"github.com/ooni/probe-engine/model"
 	"github.com/ooni/probe-engine/netx/archival"
 	"github.com/ooni/probe-engine/netx/modelx"
 	"github.com/ooni/probe-engine/netx/trace"
+	"github.com/ooni/probe-engine/resources"
 )
 
 func TestNewTCPConnectList(t *testing.T) {
