+++ netx/dialer/errorwrapper_test.go
@@ -6,8 +6,8 @@ import (
 	"io"
 	"testing"
 
+	"github.com/ooni/probe-engine/legacy/netx/dialid"
 	"github.com/ooni/probe-engine/netx/dialer"
-	"github.com/ooni/probe-engine/netx/internal/dialid"
 	"github.com/ooni/probe-engine/netx/modelx"
 )
